third_party/blink/web_tests/html/tabular_data/table_rows-expected.txt

Test behavior of the HTMLTableElement rows attribute in cases where there is unusual nesting.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS checkRowNesting("col") is 0
PASS checkRowNesting("colgroup") is 0
PASS checkRowNesting("div") is 0
PASS checkRowNesting("form") is 0
PASS checkRowNesting("script") is 0
PASS checkRowNesting("table") is 0
PASS checkRowNesting("td") is 0
PASS checkRowNesting("th") is 0


PASS checkRowNesting("tbody") is 0
PASS checkRowNesting("tfoot") is 0
PASS checkRowNesting("thead") is 0


PASS checkRowNesting("tr") is 1


PASS checkNoBodyRowNesting("col") is 0
PASS checkNoBodyRowNesting("colgroup") is 0
PASS checkNoBodyRowNesting("div") is 0
PASS checkNoBodyRowNesting("form") is 0
PASS checkNoBodyRowNesting("script") is 0
PASS checkNoBodyRowNesting("table") is 0
PASS checkNoBodyRowNesting("td") is 0
PASS checkNoBodyRowNesting("th") is 0


PASS checkNoBodyRowNesting("tbody") is 1
PASS checkNoBodyRowNesting("tfoot") is 1
PASS checkNoBodyRowNesting("thead") is 1


PASS checkNoBodyRowNesting("tr") is 1


PASS successfullyParsed is true

TEST COMPLETE
